requirements.txt
pandas
pyyaml
requests
streamlit 
openpyxl